Entregables/ana_karen_gama_velasco/APPIUM/AppiumAuto/requirements.txt
behave==1.2.6
Appium-Python-Client==2.8.1
python-dotenv==1.0.0
selenium==4.8.2